refactor(import): extract generic OAuth from the OneNote desktop sign-in
The loopback-redirect sign-in mixed three concerns. Separate them so the OAuth
machinery is reusable for any future provider:

- services/oauth/oauth.ts: provider-agnostic public-client PKCE OAuth
  (PKCE, authorize URL, code exchange, refresh), driven by an OAuthProviderConfig.
- desktop/services/loopback_oauth.ts: generic authorizeViaLoopback(config) —
  loopback server + browser + exchange, with no provider knowledge.
- import/onenote/oauth.ts: now just the OneNote OAuthProviderConfig (the single
  source of provider info).
- desktop/services/onenote.ts + the API route: thin consumers that pass the
  OneNote config to the generic helpers.

No behaviour change; the web flow and the desktop flow are unchanged.
